Shortened file paths for modules to avoid exceeding limits on Windows.
The issue was more obvious when using SOURCE_DIR with CPM_AddModule, since the absolute path was used multiple times, generating long paths.
On top of that, Visual C++ generated even longer paths, causing builds to fail.

This solution changes the way paths to modules are generated; instead of cleaning the paths, they are instead hashed.